read decimals encoded as fixed width byte arrays
This adds support for reading decimal data types encoded as fixed width byte arrays.
Fields must be annotated as Decimal logical types. The transformed type is decided based on the scale and precision of the decimal annotation.

The mapping is done automatically when annotation is present, but a mapper function is also provided that can be used for additional control and in cases where there is no annotation.

`logical_decimal(v, precision, scale; use_float=true)`

This function reads decimals from fixed length byte array fields. This converts the data read as fixed length `Vector{UInt8}` types to `Integer`, `Float64` or `Decimal` of the given precision and scale, depending on the options provided.

- `scale == 0`
    - `precision < 5`: `Int16`
    - `precision < 10`: `Int32`
    - `precision < 19`: `Int64`
    - `precision >= 19`: `Int128`
- `scale != 0`
    - `Decimal` by default
    - `Float64` when the `logical_decimal` function is used as custom mapping function with `use_float` set to `true`